requirements.lock.txt
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cli_helpers==2.7.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==45.0.5
curl_cffi==0.13.0
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.116.1
fonttools==4.59.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
multitasking==0.0.12
narwhals==2.0.1
numba==0.61.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
passlib==1.7.4
peewee==3.18.2
pgcli==4.3.0
pgspecial==2.2.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.2
streamlit==1.49.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
truststore==0.10.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xlsxwriter==3.2.5
yfinance==0.2.65
